Preserve image type field, add size, cams ns.
Size is :thumbnail if the image type is "T" or "I", and :small, :medium,
or :large depending on dimensions otherwise.

Also, I must have been going crazy; reversing the images seemed wrong
when I was doing it, and now appears to be wrong. Probably I should add
some real tests.