Don't rescrape everything if our last img is gone
Before, we just used the id of the most recent image we'd seen in order
to determine where to stop when fetching new images. It seems that
sometimes, images are published and later removed, which would cause us
to rescrape everything, since we'd never run into our old image. Now, we
use the 100 most recent images to figure out when to stop, so we'll be
fine unless they remove 100 or more images that we've already scraped.